BiocManager-pkg {BiocManager}	R Documentation

Install or update Bioconductor, CRAN, or GitHub packages

Description

This package provides tools for managing Bioconductor and other packages in a manner consistent with Bioconductor's package versioning and release system.

Details

Main functions are as follows; additional help is available for each function, e.g., ?BiocManager::version.

BiocManager::install()

Install or update packages from Bioconductor, CRAN, and GitHub.

BiocManager::version()

Report the version of Bioconductor in use.

BiocManager::available()

Return a character() vector of package names available (at BiocManager::repositories()) for installation.

BiocManager::valid()

Determine whether installed packages are from the same version of Bioconductor.

BiocManager::repositories()

Bioconductor and other repository URLs to discover packages for installation.

Author(s)

Maintainer: Marcel Ramos marcel.ramos@roswellpark.org (0000-0002-3242-0582) [contributor]

Authors:

Martin Morgan martin.morgan@roswellpark.org (0000-0002-5874-8148)

See Also

Useful links:

Report bugs at https://github.com/Bioconductor/BiocManager/issues

[Package BiocManager version 1.30.4 Index]