version {BiocManager}	R Documentation

Version of Bioconductor currently in use.

Description

version() reports the version of Bioconductor appropropriate for this version of R, or the version of Bioconductor requested by the user.

Usage

version()

## S3 method for class 'version_sentinel'
print(x, ...)

Arguments

x
An unknown_version instance used to represent the situation when the version of Bioconductor in use cannot be determined.

...
Additional arguments, ignored.

Details

version() (and all functions requiring version information) fails when version cannot be validated e.g., because internet access is not available.

Value

A two-digit version, e.g., 3.8, of class package_version describing the version of Bioconductor in use.

Examples

BiocManager::version()

[Package BiocManager version 1.30.19 Index]