BiocManager-pkg {BiocManager}	R Documentation

Install or update Bioconductor, CRAN, or GitHub packages

Description

This package provides tools for managing Bioconductor and other packages in a manner consistent with Bioconductor's package versioning and release system.

Details

Main functions are as follows; additional help is available for each function, e.g., ?BiocManager::version.

BiocManager::install(): Install or update packages from Bioconductor, CRAN, and GitHub.

BiocManager::version(): Report the version of Bioconductor in use.

BiocManager::available(): Return a character() vector of package names available (at BiocManager::repositories()) for installation.

BiocManager::valid(): Determine whether installed packages are from the same version of Bioconductor.

BiocManager::repositories(): Bioconductor and other repository URLs to discover packages for installation.

The version of Bioconductor in use is determined by the installed version of a second package, BiocVersion. BiocVersion is installed automatically during first use of BiocManager::install(). If BiocVersion has not yet been installed, the version is determined by code in base R.

Options influencing package behavior (see ?options, ?getOption) include:

"repos", "BiocManager.check_repositories", "BiocManager.snapshot": URLs of additional repositories for use by BiocManger::install(). See ?repositories.

"pkgType": The default type of packages to be downloaded and installed; see ?install.packages.

"timeout": The maximum time allowed for download of a single package, in seconds. BiocManager increases this to 300 seconds to accommodate download of large BSgenome and other packages.

System environment variables influencing package behavior include:

BIOCONDUCTOR_ONLINE_VERSION_DIAGNOSIS advanced configuration to avoid Bioconductor version checks. See ?install.

Author(s)

Maintainer: Marcel Ramos marcel.ramos@roswellpark.org (ORCID) [contributor]

Authors:

Martin Morgan martin.morgan@roswellpark.org (ORCID)

See Also

Useful links:

Report bugs at https://github.com/Bioconductor/BiocManager/issues

Examples

R.version.string
packageVersion("BiocManager")
if (requireNamespace("BiocVersion", quietly = TRUE))
    packageVersion("BiocVersion")
BiocManager::version()

[Package BiocManager version 1.30.19 Index]