available {BiocManager}	R Documentation

Discover packages available for installation.

Description

The function lists all packages available from repositories() when no pattern is provided. This usually includes CRAN and Bioconductor packages. The function can also be used to check for package name availability. Common use cases include annotation package lookups by organism short name (e.g., "hsapiens").

Usage

available(pattern = "", include_installed = TRUE)

Arguments

pattern
character(1) pattern to filter (via grep(pattern=...)) available packages; the filter is not case sensitive.

include_installed
logical(1) When TRUE, include installed packages in list of available packages; when FALSE, exclude installed packages.

Value

character() vector of package names available for installation.

Examples

if (interactive()) {
  avail <- BiocManager::available()
  length(avail)

  BiocManager::available("bs.*hsapiens")
}

[Package BiocManager version 1.30.19 Index]